requirements.txt
python_docx==0.8.11
pyttsx3==2.90